Fix Dictionary.word_list() file name path bug
There was a bug in the word_list() function  when using the Dictionary
app in the Hangman app. The file path was not set up correctly so when
used in Hangman, an error was returned.

I changed the file path name to use "../assets/words.txt" and added `|>
Path.expand(__DIR__)`. This enables the Hangman app to be able to
access the words list in the Dictionary app.