Add test for game and clear hangman test
The hangman_test was the default test that was generated with the app
and did not contain anything that tested the app. I removed the test.

I created game_test and added a basic test to ensure that when
`game_new()` is ran, we get back our expected number of turns,
game_state, and letters.